Fix: 결제 모듈 통합 및 프로젝트 상태 업데이트 기능 추가
- ProjectDetailMapper에 updateProjectStatus 메서드 추가
- PaymentServiceImpl에서 결제 완료 시 프로젝트 상태를 IN_PROGRESS로 변경
- ContractMilestoneVO, ContractMilestoneResponseDTO의 milestoneId 타입을 Long에서 Integer로 변경
- ContractMilestoneMapper에 selectMilestoneById, updateMilestoneStatusById 메서드 추가
- ProjectDetailMapper.xml의 중복된 updateProjectStatus 정의 제거
- XML 파싱 오류 및 중복 정의 오류 해결
- ContractService에 requestPayment 메서드 추가
- 사용하지 않는 applier.jsp 파일 삭제

+++ src/main/java/com/sanaiclub/contract/model/dto/ContractMilestoneResponseDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ContractMilestoneResponseDTO {
-    private Long milestoneId;
+    private Integer milestoneId;
     private Integer step;
     private String title;
     private String description;
